refactor: module restructures
